Refactor to python_tado_ha
