+++ src/python_tado_ha/const.py
