+++ src/python_tado_ha/exceptions.py
