+++ src/python_tado_ha/models.py
